压力板失效延迟改为float类型，默认值0.4

+++ Assets/Scripts/机制配置.cs
@@ -7,7 +7,7 @@ using UnityEngine;
 public class 机制配置 : SerializedScriptableObject
 {
 
-    public int 压力板失效延迟 = 1;
+    public float 压力板失效延迟 = 1;
 
     [LabelText("藤蔓茎生长(上下)")]
     public 藤蔓.生长 藤蔓茎生长;
